fix: seperate out model-watch functions for multi controller compat
