+++ src/lib/rico-angular.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, ApplicationRef } from '@angular/core';
 import { getService, LoggerFactory } from 'rico-js';
 import { ControllerProxy } from './controller-proxy';
+import { ModelMaintainer } from './model-maintainer';
 
 @Injectable({
   providedIn: 'root'
@@ -10,20 +11,21 @@ export class RicoService {
 
   private contextFactory: any;
   private clientContext: any;
-  private appRef: ApplicationRef;
+  private modelMaintainer: ModelMaintainer;
 
-  constructor() {
+  constructor(modelMaintainer: ModelMaintainer) {
     RicoService.LOGGER.debug('RicoService created');
+    this.modelMaintainer = modelMaintainer;
   }
 
   connect(remotingEndpoint: string, appRef: ApplicationRef): Promise<any> {
-    this.appRef = appRef;
     if (!this.contextFactory) {
       this.contextFactory = this.getClientContextFactory();
     }
 
     if (!this.clientContext) {
       this.clientContext = this.contextFactory.create(remotingEndpoint);
+      this.modelMaintainer.init(appRef, this.clientContext);
     }
 
     return this.clientContext.connect();
@@ -43,7 +45,7 @@ export class RicoService {
   }
 
   getHttpClient() {
-    return getService('HttpClilent');
+    return getService('HttpClient');
   }
 
   getLogger(name: string) {
@@ -52,7 +54,7 @@ export class RicoService {
 
   createController(name: string): Promise<ControllerProxy> {
     if (this.clientContext && this.clientContext.isConnected) {
-      const controllerProxy = new ControllerProxy(this.appRef, this.clientContext);
+      const controllerProxy = new ControllerProxy(this.clientContext);
       return controllerProxy.create(name);
     } else {
       return Promise.reject('Cannot create controller. ClientContext not conntected');
